+++ changelogs/fragments/vm_fix.yml
@@ -1,2 +1,6 @@
 bugfixes:
-    - fix issue with incorrect dict update in vars manager
+    - fix FactCache.update() to conform to the dict API.
+minor_changes:
+    - Moved the FactCache code from ansible.plugins.cache.FactCache to
+      ansible.vars.fact_cache.FactCache as it is not meant to be used to
+      implement cache plugins.
